# Improved performance when decoding the entire set of rows with streamable JSON formats