--- a/benchmarks/common/index.ts
+++ b/benchmarks/common/index.ts
@@ -1,0 +1,1 @@
+export * from './handlers'