--- a/benchmarks/leaks/memory_leak_arrays.ts
+++ b/benchmarks/leaks/memory_leak_arrays.ts
@@ -1,7 +1,8 @@
-import { v4 as uuid_v4 } from 'uuid'
+import { createClient } from '@clickhouse/client'
 import { randomInt } from 'crypto'
+import { v4 as uuid_v4 } from 'uuid'
+import { attachExceptionHandlers } from '../common'
 import {
-  attachExceptionHandlers,
   getMemoryUsageInMegabytes,
   logFinalMemoryUsage,
   logMemoryUsage,
@@ -9,7 +10,6 @@ import {
   randomArray,
   randomStr,
 } from './shared'
-import { createClient } from '@clickhouse/client'
 
 const program = async () => {
   const client = createClient({})