--- a/benchmarks/leaks/shared.ts
+++ b/benchmarks/leaks/shared.ts
@@ -1,15 +1,5 @@
 import { memoryUsage } from 'process'
 
-export function attachExceptionHandlers() {
-  process.on('uncaughtException', (err) => logAndQuit(err))
-  process.on('unhandledRejection', (err) => logAndQuit(err))
-
-  function logAndQuit(err: unknown) {
-    console.error(err)
-    process.exit(1)
-  }
-}
-
 export interface MemoryUsage {
   rss: number
   heapTotal: number