--- a/packages/client-common/src/index.ts
+++ b/packages/client-common/src/index.ts
@@ -44,12 +44,22 @@ export {
 } from './settings'
 
 /** For implementations usage only - should not be re-exported */
+export type {
+  RawDataFormat,
+  JSONDataFormat,
+  StreamableDataFormat,
+  StreamableJSONDataFormat,
+  SingleDocumentJSONFormat,
+} from './data_formatter'
 export {
+  formatQuerySettings,
+  formatQueryParams,
   encodeJSON,
   isSupportedRawFormat,
+  isStreamableJSONFamily,
+  isNotStreamableJSONFamily,
   decode,
   validateStreamFormat,
-  StreamableDataFormat,
 } from './data_formatter'
 export {
   type ValuesEncoder,
@@ -84,11 +94,5 @@ export type {
   ConnPingResult,
   ConnOperation,
 } from './connection'
-export {
-  type RawDataFormat,
-  type JSONDataFormat,
-  formatQuerySettings,
-  formatQueryParams,
-} from './data_formatter'
 export type { QueryParamsWithFormat } from './client'
 export type { IsSame } from './ts_utils'