--- a/packages/client-common/src/index.ts
+++ b/packages/client-common/src/index.ts
@@ -58,7 +58,6 @@ export {
   isSupportedRawFormat,
   isStreamableJSONFamily,
   isNotStreamableJSONFamily,
-  decode,
   validateStreamFormat,
 } from './data_formatter'
 export {