## Summary

* Improved performance when decoding the entire set of rows with _streamable_ JSON formats (such as `JSONEachRow` or `JSONCompactEachRow`) by calling the `ResultSet.json()` method. Depending on the dataset, it's between 10-15% (like [cell_towers](https://clickhouse.com/docs/en/getting-started/example-datasets/cell-towers)) and 40% (with large rows and very long strings) less execution time. NB: The actual streaming performance when consuming the `ResultSet.stream()` (which was fast) hasn't changed. Only the `ResultSet.json()` method used a suboptimal stream processing in some instances, and now `ResultSet.json()` just consumes the same stream transformer provided by the `ResultSet.stream()` method.

  Before:
  ```
  [1277 ms   ][JSONEachRow       ][SELECT * FROM large_strings ORDER BY id ASC LIMIT 100000                        ]
  [668 ms    ][JSONEachRow       ][SELECT * FROM cell_towers ORDER BY (radio, mcc, net, created) ASC LIMIT 500000  ]
  ```
  After:
  ```
  [737 ms    ][JSONEachRow       ][SELECT * FROM large_strings ORDER BY id ASC LIMIT 100000                        ]
  [578 ms    ][JSONEachRow       ][SELECT * FROM cell_towers ORDER BY (radio, mcc, net, created) ASC LIMIT 500000  ]
  ```

* Removed the outdated `decode` function.
* Updated exported types and doc entries for DataFormat.
* Fixed weird flakiness in expect-type assertions