@slvrtrn 2024-03-29
@mshustov 
> What change gave the performance boost? 
The `decode` function did a few unneeded passes on a huge string. It is way less efficient than the stream transformer we already had in the ResultSet.