@slvrtrn 2024-03-29
This: https://github.com/ClickHouse/clickhouse-js/pull/253/files#diff-46e9eda9880b41c04095d911536817d628992924b2bdd3a96f66529dd3628a5bL101-L104

Unless it's optimized out by JIT, it's at least one extra full pass, worst case two.